+++ tests/test_utils.py
@@ -412,54 +412,3 @@ class BMIForcingFixture_HistoricalRegrid(BMIForcingFixture):
         ]
         # Update test fixture status
         self._state = "post_ran"
-
-
-@pytest.fixture
-def bmi_forcing_fixture(request) -> BMIForcingFixture:
-    """Constructor for minimal class of classes for running BMI forcing.
-    For example usage, see: tests/esmf_regrid/test_esmf_regrid.test_regrid_aorc_aws.
-
-    Parameters:
-        request is a built-in convention for pytest.fixture.  It may be passed from @pytest.mark.parametrize usage elsewhere.
-    """
-    (config_file,) = request.param
-    bmi_model = NWMv3_Forcing_Engine_BMI_model()
-    bmi_model.initialize_with_params(
-        config_file=config_file,
-        b_date=None,
-        geogrid=None,
-        output_path=None,
-    )
-    return BMIForcingFixture(bmi_model=bmi_model)
-
-
-@pytest.fixture
-def bmi_forcing_fixture_historical_regrid(
-    request,
-) -> BMIForcingFixture_HistoricalRegrid:
-    """Constructor for minimal class of classes for running BMI historical forcing ESMF regrid functions.
-    For example usage, see: tests/esmf_regrid/test_esmf_regrid.test_regrid_aorc_aws.
-
-    Parameters:
-        request is a built-in convention for pytest.fixture.  It may be passed from @pytest.mark.parametrize usage elsewhere.
-    """
-    (
-        regrid_func,
-        config_file,
-        regrid_arrays_to_trim_extra_elements,
-        keys_to_check,
-    ) = request.param
-
-    bmi_model = NWMv3_Forcing_Engine_BMI_model()
-    bmi_model.initialize_with_params(
-        config_file=config_file,
-        b_date=None,
-        geogrid=None,
-        output_path=None,
-    )
-    return BMIForcingFixture_HistoricalRegrid(
-        bmi_model=bmi_model,
-        regrid_func=regrid_func,
-        regrid_arrays_to_trim_extra_elements=regrid_arrays_to_trim_extra_elements,
-        keys_to_check=keys_to_check,
-    )
